refactor: rename gameloop functions to increase readility
