+++ src/ui/prompt/prompt.ts
@@ -1,13 +1,4 @@
 import * as readline from 'readline'
-import { State, Color } from '../../chess/chess.models'
-import { sample } from '../../utils/array'
-import { drawState } from '../draw/draw'
-import { WHITE } from '../../chess/chess.constants'
-import { isEmpty } from '../../utils/fp'
-import { generateMoves, playMove } from '../../chess/moves/moves'
-import { isInCheck } from '../../chess/moves/moves.lib'
-import { moveToReadable } from '../draw/draw.lib'
-import logger from '../../utils/logger'
 
 export async function getInput(prompt: string): Promise<string> {
 	const rl = readline.createInterface({
@@ -21,46 +12,6 @@ export async function getInput(prompt: string): Promise<string> {
 			resolve(input)
 		})
 	})
-}
-
-export async function playTurnInConsole(state: State, userSide: Color) {
-	drawState(state, userSide)
-	const possibleMoves = generateMoves(state)
-
-	logger.debug(
-		'turn',
-		state.sideToMove === WHITE ? 'White' : 'Black',
-		'legal moves',
-		possibleMoves.map((move) => moveToReadable(state, move)),
-		'in check',
-		isInCheck(state),
-	)
-
-	if (isEmpty(possibleMoves)) {
-		const inCheck = isInCheck(state)
-		if (inCheck) {
-			logger.info(
-				state.sideToMove === WHITE
-					? 'Black won with checkmate'
-					: 'White won with checkmate',
-			)
-			return
-		}
-		logger.info('Stalemate.')
-		return
-	}
-
-	if (state.sideToMove === userSide) {
-		logger.info('Your turn.')
-		const playedMove = await getInput('Enter move: (e.g. a3a5)')
-		const newState = playMove(playedMove, state)
-		return playTurnInConsole(newState, userSide)
-	}
-
-	const moves = generateMoves(state)
-	const playedMove = sample(moves)
-	const newState = playMove(playedMove, state, true)
-	return playTurnInConsole(newState, userSide)
 }
 
 export async function promptForGameType(): Promise<'w' | 'b' | 'c'> {
@@ -78,37 +29,4 @@ export async function promptForGameType(): Promise<'w' | 'b' | 'c'> {
 
 	console.info('Invalid color.')
 	return promptForGameType()
-}
-
-export async function startComputerOnlyGame(state: State) {
-	drawState(state, WHITE)
-
-	const possibleMoves = generateMoves(state)
-
-	logger.debug(
-		'turn',
-		state.sideToMove === WHITE ? 'White' : 'Black',
-		'legal moves',
-		possibleMoves.map((move) => moveToReadable(state, move)),
-		'in check',
-		isInCheck(state),
-	)
-
-	if (isEmpty(possibleMoves)) {
-		const inCheck = isInCheck(state)
-		if (inCheck) {
-			console.info(
-				state.sideToMove === WHITE
-					? 'Black won with checkmate'
-					: 'White won with checkmate',
-			)
-			return
-		}
-		console.info('Stalemate.')
-		return
-	}
-
-	const move = sample(possibleMoves)
-	const newState = playMove(move, state, true)
-	setTimeout(() => startComputerOnlyGame(newState), 250)
 }
